+++ animation_nodes/nodes/mesh/set_bevel_edge_weight.py
@@ -25,7 +25,9 @@ class SetBevelEdgeWeightNode(AnimationNode, bpy.types.Node):
         if object.mode != "OBJECT":
             self.raiseErrorMessage("Object is not in object mode.")
 
+        attribute = object.data.attributes.get("bevel_weight_edge")
+        if not attribute: attribute = object.data.attributes.new("bevel_weight_edge", "FLOAT", "EDGE")
         weights = VirtualDoubleList.create(weights, 0).materialize(len(object.data.edges))
-        object.data.edges.foreach_set('bevel_weight', weights)
+        attribute.data.foreach_set('value', weights)
         object.data.update()
         return object